+++ forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_1_2/eventsubscription/EventSubscriptionApiController.java
@@ -24,13 +24,15 @@ import com.forgerock.openbanking.aspsp.rs.store.repository.TppRepository;
 import com.forgerock.openbanking.aspsp.rs.store.repository.v3_1_2.events.EventSubscriptionsRepository;
 import com.forgerock.openbanking.common.conf.discovery.ResourceLinkService;
 import com.forgerock.openbanking.common.model.openbanking.event.FREventSubscription1;
-import com.forgerock.openbanking.common.services.openbanking.event.EventValidationService;
+import com.forgerock.openbanking.common.model.version.OBVersion;
+import com.forgerock.openbanking.common.services.openbanking.event.EventResponseUtil;
 import com.forgerock.openbanking.exceptions.OBErrorResponseException;
 import com.forgerock.openbanking.model.Tpp;
 import com.forgerock.openbanking.model.error.OBRIErrorResponseCategory;
 import com.forgerock.openbanking.model.error.OBRIErrorType;
 import io.swagger.annotations.ApiParam;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -54,11 +56,21 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
     private final EventSubscriptionsRepository eventSubscriptionsRepository;
     private final TppRepository tppRepository;
     private final ResourceLinkService resourceLinkService;
+    private final EventResponseUtil eventResponseUtil;
 
+    @Autowired
     public EventSubscriptionApiController(EventSubscriptionsRepository eventSubscriptionsRepository, TppRepository tppRepository, ResourceLinkService resourceLinkService) {
         this.eventSubscriptionsRepository = eventSubscriptionsRepository;
         this.tppRepository = tppRepository;
         this.resourceLinkService = resourceLinkService;
+        this.eventResponseUtil = new EventResponseUtil(resourceLinkService, OBVersion.v3_1_2, true);
+    }
+
+    public EventSubscriptionApiController(EventSubscriptionsRepository eventSubscriptionsRepository, TppRepository tppRepository, ResourceLinkService resourceLinkService, EventResponseUtil eventResponseUtil) {
+        this.eventSubscriptionsRepository = eventSubscriptionsRepository;
+        this.tppRepository = tppRepository;
+        this.resourceLinkService = resourceLinkService;
+        this.eventResponseUtil = eventResponseUtil;
     }
 
     public ResponseEntity createEventSubscription(
@@ -119,7 +131,7 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
                 .body(packageResponse(frEventSubscription1));
     }
 
-    public ResponseEntity<OBEventSubscriptionsResponse1> readEventSubscription(
+    public ResponseEntity readEventSubscription(
             @ApiParam(value = "An Authorisation Token as per https://tools.ietf.org/html/rfc6750", required = true)
             @RequestHeader(value = "Authorization", required = true) String authorization,
 
@@ -134,21 +146,27 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
             Principal principal
     ) throws OBErrorResponseException {
         log.debug("Read event subscription for client: {}", clientId);
-
-        return Optional.ofNullable(tppRepository.findByClientId(clientId))
-                .map(Tpp::getId)
-                .map(eventSubscriptionsRepository::findByTppId)
-                .map(eventSubs -> ResponseEntity.status(HttpStatus.OK).body(packageResponse(eventSubs)))
-                .orElseThrow( () ->
-                            new OBErrorResponseException(
-                                HttpStatus.NOT_FOUND,
-                                OBRIErrorResponseCategory.REQUEST_INVALID,
-                                OBRIErrorType.TPP_NOT_FOUND.toOBError1(clientId)
-                            )
-                );
+        final Optional<Tpp> tppId = Optional.ofNullable(tppRepository.findByClientId(clientId));
+        if (!tppId.isEmpty()) {
+            // A TPP must not access a event-subscription on an older version, via the EventSubscriptionId for an event-subscription created in a newer version
+            Collection<FREventSubscription1> frEventSubscription1s = Optional.ofNullable(eventSubscriptionsRepository.findByTppId(tppId.get().id))
+                    .orElseGet(Collections::emptyList)
+                    .stream().filter(frEventSubs -> eventResponseUtil.IsAllowedAccessResourceFromApiVersionInstanced(frEventSubs.obEventSubscription1.getData().getVersion())).collect(Collectors.toList());
+            if (!frEventSubscription1s.isEmpty()) {
+                return ResponseEntity.status(HttpStatus.OK).body(packageResponse(frEventSubscription1s));
+            } else {
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("The event subscription can't be read via an older API version.");
+            }
+        } else {
+            throw new OBErrorResponseException(
+                    HttpStatus.NOT_FOUND,
+                    OBRIErrorResponseCategory.REQUEST_INVALID,
+                    OBRIErrorType.TPP_NOT_FOUND.toOBError1(clientId)
+            );
+        }
     }
 
-    public ResponseEntity<OBEventSubscriptionResponse1> updateEventSubscription(
+    public ResponseEntity updateEventSubscription(
             @ApiParam(value = "EventSubscriptionId", required = true)
             @PathVariable("EventSubscriptionId") String eventSubscriptionId,
 
@@ -172,7 +190,6 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
 
             Principal principal
     ) throws OBErrorResponseException {
-
         final OBEventSubscription1 updatedSubscription =
                 new OBEventSubscription1().data(new OBEventSubscription1Data()
                         .callbackUrl(obEventSubscriptionParam.getData().getCallbackUrl())
@@ -182,11 +199,14 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
         final Optional<FREventSubscription1> byId = eventSubscriptionsRepository.findById(eventSubscriptionId);
         if (byId.isPresent()) {
             FREventSubscription1 existingEventSubscription = byId.get();
-            EventValidationService.checkEqualOrNewerVersion(existingEventSubscription.getObEventSubscription1(), updatedSubscription);
-
-            existingEventSubscription.setObEventSubscription1(updatedSubscription);
-            eventSubscriptionsRepository.save(existingEventSubscription);
-            return ResponseEntity.ok(packageResponse(existingEventSubscription));
+            // A TPP must not update a event-subscription on an older version, via the EventSubscriptionId for an event-subscription created in a newer version
+            if (eventResponseUtil.IsAllowedAccessResourceFromApiVersionInstanced(existingEventSubscription.obEventSubscription1.getData().getVersion())) {
+                existingEventSubscription.setObEventSubscription1(updatedSubscription);
+                eventSubscriptionsRepository.save(existingEventSubscription);
+                return ResponseEntity.ok(packageResponse(existingEventSubscription));
+            } else {
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("The event subscription can't be update via an older API version.");
+            }
         } else {
             // PUT is only used for amending existing subscriptions
             throw new OBErrorResponseException(
@@ -216,9 +236,14 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
     ) throws OBErrorResponseException {
         final Optional<FREventSubscription1> byId = eventSubscriptionsRepository.findById(eventSubscriptionId);
         if (byId.isPresent()) {
-            log.debug("Deleting event subscriptions URL: {}", byId.get());
-            eventSubscriptionsRepository.deleteById(eventSubscriptionId);
-            return ResponseEntity.noContent().build();
+            // A TPP must not delete a event-subscription on an older version, via the EventSubscriptionId for an event-subscription created in a newer version
+            if (eventResponseUtil.IsAllowedAccessResourceFromApiVersionInstanced(byId.get().obEventSubscription1.getData().getVersion())) {
+                log.debug("Deleting event subscriptions URL: {}", byId.get());
+                eventSubscriptionsRepository.deleteById(eventSubscriptionId);
+                return ResponseEntity.noContent().build();
+            } else {
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("The event subscription can't be delete via an older API version.");
+            }
         } else {
             throw new OBErrorResponseException(
                     HttpStatus.BAD_REQUEST,
@@ -247,7 +272,7 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
                     .data(new OBEventSubscriptionsResponse1Data()
                             .eventSubscription(eventSubsByClient))
                     .meta(new Meta())
-                    .links(resourceLinkService.toSelfLink(discovery -> discovery.getV_3_1_2().getGetCallbackUrls()));
+                    .links(resourceLinkService.toSelfLink(discovery -> discovery.getVersion(eventResponseUtil.version).getGetCallbackUrls()));
         }
     }
 
@@ -260,7 +285,7 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
                         .eventTypes(obEventSubs.getEventTypes())
                         .version(obEventSubs.getVersion())
                 )
-                .links(resourceLinkService.toSelfLink(frEventSubscription1, discovery -> discovery.getV_3_1_2().getGetCallbackUrls()))
+                .links(resourceLinkService.toSelfLink(frEventSubscription1, discovery -> discovery.getVersion(eventResponseUtil.version).getGetCallbackUrls()))
                 .meta(new Meta());
     }
 }
